Move Resistor to models folder and start using module alias
Moving Resistor to a place that can be easily imported by components and
containers.

When we want to move files into different folders we don't want to keep
updating all files to the correct references, thus adding alias for path
imports (we want to keep diffs as small as possible).